Updating ez with the latest changes
There is probably not going to be a full release soon, since we are in
the middle of several larger projects.

Therefore this commit just updates the base libraries with the latest
changes. Note that this update has not been tested the way we test full
releases.

Notable changes:
* Added ezOBJLoader to load OBJ/MTL files.
* Renamed ezStringIterator to ezStringView.
* Added ezGeometry, a class with functions to generate basic geometric
shapes.
* Added ezChunkFormatReader / ezChunkFormatWriter to work with 'chunked'
file formats.
* Moved the "Foundation\Basics\Types" folder one folder up.
* Lots of minor fixes and enhancements.