Fixed race condition in 'Reallocate' in its use of ezMemoryTracker if it fails to extend the original address and has to actually reallocate. Removed ReplaceAllocation as there is never a case in which this function would be thread safe.
In case 'this->m_allocator.Reallocate' returns a different pointer, the original address is already freed and can (AND WILL) be re-used. But it is not yet removed from the memory tracker, so if another thread takes the just released memory block and registers it before 'Reallocate' can go into 'ezMemoryTracker::ReplaceAllocation' you have got yourself a race condition.