-Logging in editor is not thread safe and no longer crashes ;) -Qt messages are now forwarded to our log and show up in the editor   -Fixes various Qt errors found from that. -Log no longer resets on every message. -Asset Curator now caches entries between session (AssetCache/AssetCurator.ezCache), making startup a lot faster (e.g. 2200ms => 600ms) -Asset Curator will automatically convert a [DataDirName].ezCollectionAsset in every data dir on init. I put some files in base there so e.g. render pipelines are always transformed on startup.
-YOU BROKE ALL THE BINARIES? YOU NEED TO GET RID OF ALL THE CACHES BECAUSE YOU FUCKED UP EVERYTHING?
We have the tool for you! Just increase #define EZ_CURATOR_CACHE_VERSION 1 by a notch and the Curator will obliterate any AssetCache folder that contains an outdated cache file. To prevent any unintentional fallout from this, the curator will not remove the cache if there does not exist a cache file yet.

ANd I wanted t fix the 88 byte memory leak but someone came before me. Turns out AddAttributes(new ezDefaultValueAttribute("")), causes the leak as "" is converted to an ezVariant, making a heap alloc with EZ_DEFAULT_NEW that is tracked but of course never is release because it is part of the static init RTTI info. Wasn't there a way to prevent plugin loading to be tracked?