Fixed ezGameEngineTest
DoSetupGraphicsDevice() & DoSetupDefaultResources() need to be called manually since a recent change in the GameEngine app framework. (which was necessary for custom device via gamestate)
Todo: This caused a lot of problems in a lot of places, should consider bringing old behavior back with a configurable "DoProjectSetup" that will by default also create device and DefaultResources.