BROKE ALL VISUAL SCRIPTS !
Changed how visual scripts declare variables (which can be exposed as parameters).
Variables are now defined on the Visual Script asset root node, those properties are shown when no node is selected.
Since for that the asset document base class was changed, old scripts don't properly work anymore (and will assert during transform) BUT it is easy to recreate them:
* Open your old VS
* Create a new VS document (right click on old in asset browser -> new VS)
* Select all nodes in old VS, copy & paste over to new VS, save
* Migrate all references from old VS to new VS
* Close editor, delete old VS, rename new VS to same name

Although there is a graph patch function to change the base class, I don't think it would work, becasue the new base class changes which object is the 'root' object, and that data is already serialized in old documents.
If you want to try out writing a patch, just use visual scripts from before this revision for testing.