Proof of concept for exposed parameters on prefabs.
There is currently one hardcoded parameter "PrefabColor" that is bound to the first found MeshComponent color property in a prefab.
To enable it, you actually need to SAVE the prefab (transform is not enough), then it will appear on prefab components.

The modified value will only be applied on instantiation, which happens when stopping the scene simulation. This implementation does not support live-update of exposed parameters, but in return it can be optimized well.