It is now allowed to attach objects to node that already have prefab components.
Let's see how this works out.
To not waste memory for tracking the prefab child nodes, deactivating prefab components only works in the editor
and it will always delete all prefab spawned children, so if you attach multiple prefab components to a node,
the behavior will not be entirely correct.