CMake: Adds support for external project folders
In the CMake GUI setup "EZ_EXTERNAL_PROJECT" to point to a folder where CMakeList.txt files are that describe your project.
The folder should have the name of your project (folder name appears in the IDE) as a group.
Project should be in sub-folders. In the root CMakeList.txt of your project folder, you only need to add ez_add_all_subdirs()