Implemented a proof-of-concept for game object references in prefabs and scenes
Once a component has an ezGameObjectHandle to some object, the handle can be properly serialized and deserialized and in the case of prefabs, it will already be remapped upon instantiation to stay local to the prefab instance.
So the main problem is how to get a game object reference into the component from the editor.

I currently use the stringyfied GUID of the ezDocumentObject and pass it to the component (right click any object and select "Copy Object Reference", then paste that into the respective string property of a component).
On the runtime side, at edit time, the component can now access a delegate that can resolve the GUID to a game object handle. The component can then save that handle during export and everything is fine.
However, only references that can be resolved at edit time will work at runtime.
Ie. exposing a reference as a prefab parameter only works in the editor. At runtime, when the prefab is instantiated, it would need to resolve the stringyfied GUID, but no resolver delegate exists.
It is currently unclear how to implement that.