GameObject references that are exposed as prefab properties work outside the editor now
The ezPrefabReferenceComponent now stores ezGameObjectHandles in the binary serialized format for exposed parameters, thus enabling handle remapping.
It then encodes the remapping method into the parameter value string. A new default reference resolver function on the ezWorld is then able to decode the reference
into a regular game object handle again.

This could all be done on an ezUuid value, as well, instead of a string.