Fixes a crash at shutdown when a plugin has a link dependency on another plugin
Crash happened under these circumstances:
A game plugin had a link dependency on the particle plugin. Both were dynamically loaded into the process, however the game plugin also increased the refcount of the particle plugin due to its implicit link dependency. At shutdown due to the sorting of the ezSet, the particle plugin was always unloaded first, but due to the +1 ref-count from the game plugin, the particle plugin was actually NOT unloaded. It got only unloaded implicitly during the later unload of the game plugin.
However, that was not when the engine expected it to happen, and several systems asserted and crashed at that point.