Fixed that the Editor could not distinguish between projects with the same name
Project state (open documents etc) was stored such that when opening a project with the same name (e.g. different repository), it would try to reuse that state. Since some stuff, like open documents where stored with absolute paths, this would not work as desired.

Now all project state is prefixed with a hash of the full project path, making it unique.