Allow using an 'ezSdkRoot.txt' redirection file to point to the SDK dir
If ezEngine is hooked up as a submodule, ez apps can't find the proper folder where the SDK root is located.

With this change you can put a file 'ezSdkRoot.txt' into the top most folder of your repository. It has to contain the relative path from this directory, to the ez SDK directory.

So if your repository is in "C:\MyRepo" and ez is a submodule located in "C:\MyRepo\ezSubmodule", you would put the file into "C:\MyRepo" and it should only contain the string "ezSubmodule".

Then all tools (ezEditor etc) and all ez(Game)Applications will automatically find the correct root.