Adds a 'random seed' value to every game object
This value can be used to initialize things in components with random values, for instance decals use this now to randomly select one of the provided decal assets.

The random seed is either randomly chosen at game object creation time, or it can be fixed (a stable 'random' number). The latter is done by the editor, meaning that all manually placed objects will use a stable seed. Therefore, every decal that was manually placed will use the exact same variation no matter how often the scene is modified or exported.
The former is done for all dynamically created objects, so everything that spawns at runtime will vary truly randomly.

The random seed is propagated through prefabs, so it is now possible that components that are spawned through prefabs vary, but still in a deterministic way, ie. they will be equally stable, if the prefab was instantiated in the editor.