Adds a console to the CVars window (Editor & ezInspector) (#469)
The console can be used the same way as the in-game Quake-style console:

    Hit TAB to auto-complete
    Enter to run a command
    Arrow up/down for input history
    F2, F3 to repeat last and second-to-last command

Commands are sent to the engine process to be executed there. Auto-complete as well. That means all CVars and console functions that are known to the engine process are available, not only the ones known by the local process.

Fixes #224