Bounding boxes are now properly computed and updated for animated meshes
During mesh transform, the maximum distance of any vertex to its influencing bones is computed (a kind of skin thickness). This is a conversative value by which to grow a skeleton's bounding box, such that the mesh is fully inside.
The bounding box is now computed from the animation poses on the fly, which makes it work with all kinds of weird meshes + ragdolls.
If the new bbox is larger than the previous one, the spatial system is informed.
Every 1024 frames the bbox is reset, so that if it becomes very large due to some animation, it will shrink again later.