Added 'Export Project' functionality (#591)
Under 'Editor > Export Project' there is now an option to export all data and binaries of a project to a folder.
This is an early version and may need additional work in the future.

Which files get exported is controlled by two config files "ProjectData.ezExportFilter" and "ProjectBinaries.ezExportFilter". These are simple path filter rules, similar to .gitignore files. There are base config files under "Data\Base" that get automatically pulled in. See those for inspiration (documentation coming later).

For all found scenes in a project, launch.bat files are generated to start the Player with that scene. Some scenes are known to currently not start (e.g. the Corridor scene). The reason is yet unknown. Also the .bat files currently have to use absolute paths to the scenes, so are not relocatable. This will be fixed in the near future.

Related: #590