Merged PR 7457859: Add support for NV12
Add support for NV12.

This adds support for loading, saving, and converting NV12 textures. Useful for investigations around video encoding. Slightly complicated since previously there was no support for planar formats at all, but the basics are added now.

Besides mip, face and array index, viewing subparts of an image now also supports supplying a plane index. In terms of parameter ordering this should be specified between array index and z slice, but for backwards compat the plane index is added as the last parameter. Viewing a single plane changes the format of the view (plane 0 of NV12 is viewed as R8_UNORM and plane 1 as R8G8_UNORM) as well as the dimensions of the view so that a subplane can be fully treated as if it was a normal image.

Code for computing block sizes, image sizes, etc. has been adjusted to support plane indices where meaningful.