Animation controller: non-looped animations will now hold the last frame until they are deactivated
This change makes it possible that a non-looped animation is played start to finish and then stays at the last frame indefinitely, as long as the anim graph node keeps being set to 'active'. This is useful for animations like "falling to the ground", where you want a character to remain in the last pose.
The previous behavior was that the node would deactivate itself, even though it is triggered, and usually that would mean that the node would be triggered again next frame and effectively loop the animation.