First ever implementation of an AudioSystem
This commit introduce the first version of an AudioSystem. The AudioSystem is a plugin abstracting most of the features provided by audio engines like Wwise, FMOD or Amplitude. But actually, any sound engine can be used as middleware, even really small ones, like SoLoud.

This commit adds a working version of the AudioSystem with support of triggering audio events from the middleware. 

The AudioSystem is splitted in a runtime plugin (ezAudioSystemPlugin), an editor plugin (ezEditorPluginAudioSystem), and a static library (ezSharedPluginAudioSystem) which will be used by the middleware editor plugin to implement required editor actions (eg: reload controls).