Asset types can now be marked as compatible with other types
This allows to add an asset type that is set up differently (different kind of document and output generation), but ultimately produces the same output and thus can be plugged into the same existing component slots without having to adjust those. This extends what plugins can add to the editor. For example a new asset type could generate a texture procedurally, but since it would write the same output file format, it can be used as a regular 2D texture in the editor.