Unlock shadow view mutex to prevent deadlocks (#661)
When many prefabs are opened at editor startup, all views try to create a shadow view. For that they try to enter this mutex on worker threads and will block the worker threads, without the tasksystem being aware of this. One thread enters the mutex and then attempts to load a pipeline, which can't be done on the same thread. However, all workers are blocked by the shadow mutex and thus no worker thread is available and we have a deadlock. The fix is to not be in the mutex while the pipeline gets loaded.