Convex mesh assets only use a single surface (#743) (#744)
Fixes #743.

Convex meshes by design only have a single surface. So far, the convex mesh asset would still extract all the material information from a source mesh and present that to users. This was a problem, because the physics engine asserted that convex meshes only have a single surface.

Now the material list is hidden and instead just a single surface property is displayed.