Added camera shake functionality (#841)
Added two components:
One to add a shake effect to a game object. This should be attached to a parent of a player camera. It injects a random deviation away from the X axis.
Another to define a volume (so far only a sphere) where camera shake should be applied. This can be placed in a level, or attached to a grenade, etc. It specifies how strong the shake effect should be.
This way, either a level or a dynamic object can add shake effects, and if the camera is close enough, it will get that shaking applied. If multiple volumes overlap, the maximum strength is used.