Fail prefab export when C++ types are missing (#874)
Currently when a prefab is exported, it will just skip unknown component types and the export succeeds.

So for example, if a game plugin is currently not compiled, and vital components are missing, the export doesn't care and just removes those components from the exported scene.

This is obviously problematic, because then those prefabs don't work as intended. It is also a problem, because the background asset processor will make sure that all assets get exported, so if you open a project that uses a C++ game plugin, it is unlikely that you manage to compile that, before all prefabs are exported in a broken state.

The solution is to detect this case, and fail the export step, so that you are forced to compile the necessary code first. Additionally, whenever the engine process is reloaded, all previously failed assets are re-evaluated, so that they automatically export successfully, once this is possible.