Fixes incorrect order of operations during scene export with thumbnail generation
Fixes #880

The scene was cleared, then the thumbnail creation was requested, then the scene was sent over. That's because the m_EngineStatus wasn't reset to Initializing and therefore the thumbnail creation message was not delayed until after the scene description was sent over.

That resultet in a skylight component being added by the thumbnail code, and then again by the scene. If the order were the other way round, the thumbnail code wouldn't create another skylight component.