Prefabs may now attach components to their parent instance node (#927)
Fixes #844:

In a prefab document, when there is only one top-most game object, you can name it "<Prefab-Root>" to turn it into a proxy object for the game object on which the prefab gets instantiated.
This makes it possible to attach components directly to the game object in a scene on which the Prefab Reference Component is attached. This is particularly useful for physics objects, since you can then have the physics components on the right node, so that prefabs can, for example, be joined together.