Fixed #230: Surfaces can now configure exposed prefab parameters
This feature was already implemented, but not active yet, since the asset system would previously deadlock in a circular dependency chain. Since the asset system was improved, this does not happen anymore.

With this change, surfaces can now edit the exposed parameters of referenced prefabs. Thus if you have an effect to spawn, e.g. a bullet hole, you can now reuse that effect for different surface types, but tweak it differently, such that for example it appears grey on a metal surface and brown on a wooden surface.