Reorganized menu structure and added support for global mapping names (#1018)
I've added many comments to ActionManager.h and ActionMap.h. For details read those.

In the editor I renamed the "Editor" menu to "Project" and moved several menu entries around.

The action system was slightly extended, it is now possible to map an action into a category, without knowing the full path to the category. You only need the category name itself, and the full path is searched for automatically. This makes it possible to use unique names for important categories and then easily map actions into them. If the category is ever moved around (or even in a different location in some context), the action automatically moves with it.

This makes it both easier to add actions to the right place, and allows us to restructure things in the future.