Migrate code to use Make* functions (#1022)
This changes all EZ code to make use of the new Make functions. The old functions have not yet been removed, so existing user code continues to compile with this change, but internally EZ makes no use of the old functions anymore.

Since the [[deprecated]] attribute is utter rubbish with MSVC (with warnings as errors you just get errors, you can't disable that), I probably won't enable it, but rather the next PR will actually remove the old functions. Though that is yet up for discussion.