Added functionality for runtime navmesh generation (#1043)
This PR adds a new plugin "AiPlugin" which adds AI functionality.

For now this is mostly about generating navmeshes at runtime (from the physics representation). And to do path searches. There is also some preliminary work for steering, but this isn't accessible from the editor, yet.

In the editor, first enable the AI plugin for your project.
Then go to Project > Plugins > AI plugin settings, to add a navmesh type and a path search type.
These define how a navmesh is generated, and which costs a character would have to traverse the different ground types of a navmesh.

Which ground type (stone, mud, water, etc) is generated from a physical object, is defined in the surfaces. Which ground types your game has, is defined in the AI settings dialog.

When you run a scene and start a path search, the system queries the physics world and builds a navmesh from that geometry. As long as no path searches are done, nothing gets generated. Generation is done in tiles, so only the necessary parts of the world get generated.

There is a new "Test NavMesh Path" component, which can be used to ask for a path and visualize the result. This is mainly for demonstrative and testing purposes. When you build your own AI, you can copy that code to do your own path queries.

There is also a new CVar "AI.Navmesh.Visualize" which lets you enable visualization for one of the navmeshes (by index). Note that nothing will show up, as long as no path was every queried, since then now mesh is generated.

![grafik](https://github.com/ezEngine/ezEngine/assets/6001174/735606f7-9b1d-4d10-abf6-6734702b5d76)