Fix pasting objects in the scene tree pastes twice
Qt's QEvent::ShortcutOverride pattern works like this:
A: If inside `QEvent::ShortcutOverride` you accept the event, it will be fired again as a regular `QEvent::keyPressEvent` to the widget the accepted the override.
B: Inside `QEvent::keyPressEvent` can execute the actual action that you requested the override for.

In our case, this means that we have to call TriggerDocumentAction twice, once to determine if we want to override and then a second time to execute the action.
Fixes #991