Added JoltClothSheetComponent (#1140)
This component uses the Jolt soft-body feature to simulate a piece of cloth, similar to the generic ClothSheetComponent. However, it interacts with physics geometry.

Use cases for this are somewhat limited, as the collision with static geometry isn't solid, and collisions with dynamic objects are even more unreliable. Additionally, it is harder to optimize away the performance impact of the simulation, since it isn't purely visual anymore. On the other hand, the performance of the Jolt simulation is amazingly fast.

This is mainly to try out the soft body functionality, and maybe build on top of it in the future.

![image](https://github.com/ezEngine/ezEngine/assets/6001174/f428c624-f9ac-498c-b7f5-23f9de2bccd8)