Fixed #1146: Added physics collision geometry visualization (#1150)
 Adds these CVars:
1. Jolt.Visualize.Geometry
2. Jolt.Visualize.Exclusive
3. Jolt.Visualize.Distance

Variable 1. enables visualization of physics collider geometry:

![image](https://github.com/ezEngine/ezEngine/assets/6001174/408e71f8-6e5b-4445-a15a-e37da6d435a0)

This is just rendered together with the regular geometry, which helps to figure out whether the physics geometry fits the render geometry. For example here the tree collision mesh is not imported the same way as the graphics mesh:

![image](https://github.com/ezEngine/ezEngine/assets/6001174/371c5f8a-6815-47e0-a31f-3a591dd69eca)

This gives a lot of z-fighting, if the meshes are identical and can be distracting. So variable 2 allows to disable rendering of regular geometry:

![image](https://github.com/ezEngine/ezEngine/assets/6001174/8fc32a6e-ba22-4f4b-b72c-ad0dab40640c)

Different types use different colors. In the image above:
* light blue = static geometry
* dark blue = kinematic
* yellow = dynamic bodies
* green = query shapes (hitboxes)
* pink = ragdolls + ropes
* red = soft bodies
* transparent = triggers

Geometry is only updated in a certain radius around the camera (to reduce the performance impact). Variable 3 configures this distance. Overall the rendering is very efficient, though, since it simply uses the CustomMeshComponent to add meshes to the scene and thus the typical rendering and culling is done on them, this is not using the ezDebugRenderer.