Changed ezStringView to work around an MSVC code generation bug
Under very specific circumstances the compile time evaluation of ezStringView got broken and it stored incorrect content.
This could only happen because ezStringView stores two pointers, rather than a start pointer and a length, as other implementations do.
To work around this, the ezStringView implementation was modified to also store a start pointer and a byte count.