Added AI navigation component and sample scene (#1225)
Adds a new ezAiNavigationComponent that can be used to have a character move along a path using the navmesh.

The component is still quite basic and to be considered work-in-progress. It will automatically search a path and navigate along it, it also accelerates and deccelerates the character and makes it round sharp corners in a more convincing manner (by slowing down). It places the character on the physical ground independent on how the navmesh looks, so that it doesn't hover in the air.

It does NOT do any avoidance of other characters or dynamic objects. This may be added later.