FileServe tool can now remote compile shaders (#1258)
Some devices can't compile shaders themselves. For example on Android we need the HLSL shaders to already be compiled for SPIR-V and can't do this on-device on-demand.

This is inconvenient, and when using FileServe to read data from a PC, it wouldn't allow to modify and hot-reload shaders. It also requires to precompile all shaders up front before launching the app.

This PR adds on-demand shader compilation through FileServe. Whenever the app needs a shader, it checks whether the shader needs to be compiled and if so, requests from FileServe to do that. FileServe then executes the shader compiler and stores the result locally on the PC. Afterwards the app attempts to read the result again, but now will find the (updated) compiled shader file through FileServe and continues normally.

Shader compilation activity shows up in the FileServe activity log, and in case of an error, the log is displayed both in the FileServe GUI and also on the remove device log.