Disable Tracy and fix some other issues (#1275)
Tracy always loads the process symbols. This locks the PDB files, which breaks the EZ engine process hot reload of the editor.

Since this is a core feature of the editor, this is unacceptable.

I've attempted to disable all Tracy callstack functionality, for which it even has options, but that doesn't prevent Tracy from locking the PDBs. Consequently, I'm disabling Tracy by default now. If somebody wants Tracy support, they can enable it manually through CMake.

Feel free to attempt to fix this, so that Tracy can be enabled by default.