SetWorldSimulation will now be cached and is only flushed into the actual simulation state in the beginning/end of ezWorld::Update() calls.
This ensures consistency for the Update-functions which only execute when the simulation is active - either all of them are run, or none.
Its no longer possible to get an in between state where only some of them are run if the simulation state is changed during an ezWorld::Update() call.