Adding parent mapping support for action maps
Previously, it was impossible to make a plugin that adds an action to every editor as you would manually have to add the action to every editor's mappings.
To fix this, inheritance has been introduced into the action mapping system.
When registering an action via `ezActionMapManager::RegisterActionMap` a  parent mapping can be provided. Once a mapping is mapped to a view, all actions of the entire parent hierarchy as show in the view.

The `ezActionMap` had to be changed slightly to account for this:
1. `MapAction` just accumulates the mappings in an array until `BuildActionTree` is called which builds the hierarchy of the inherited and local actions. This is necessary as mappings can be defined in any order so the parent pamming might not be defined yet and mapping an action into the parent mapping's hierarchy might not be possible yet. 
2. `UnmapAction` has been removed from the
public interface and replaced with `HideAction` which practically does the same, but later at the `BuildActionTree` phase.
3. `SearchPathForAction` has been made private as is requires the hierarchy to be built already which is not possible when defining the mapping across parent hierarchies which are not defined yet. The few use cases we had have been fixed by adding a parameter `const ezBitflags<ezStandardMenuTypes> menus` to `ezProjectActions::MapActions` to allow for partial mapping of actions.

Right now, the following base action mappings have been defined which ca nnow be extended via plugins to easily map to every asset document editor's UI:
1. **AssetMenuBar**: The bare minimum of actions every asset document's menu bar should have.
2. **AssetToolbar**: The bare minimum of actions every asset document's toolbar should have.
3. **AssetViewToolbar**: Empty but should be the base mapping of every view.
4. **SimpleAssetViewToolbar**: Convenience mapping that derived from *AssetViewToolbar* and provides the most common actions used by engine views.