Generalized support for injecting textures into a render pipeline (#1319)
* Generalized support for injecting textures into a render pipeline

Previously, the only input to a render pipeline was the `ezTargetPass` which was hardcoded in the `ezRenderPipeline`. The concept has now been generalized with these changes:
1. New node pins `ezRenderPipelineNodeOutputProviderPin` and `ezRenderPipelineNodeInputProviderPin` have been added.
2. When the new pin type is present, the render pipeline asks the node for a texture to be provided for this pin by calling `ezRenderPipelinePass::QueryTextureProvider` with the pin and texture description that is expected.
3. If the pass returns a valid texture, this texture is used instead for the connection instead of using the texture pool. In case an invalid texture is returned, the pin reverts to the normal input / output variant.
4. The hard-coded logic has been moved from the render pipeline to the `ezTargetPass::QueryTextureProvider` implementation.

New render passes have been added that built on top of the new feature and allows to access data from a previous frame. To preserve textures across the next frame you need to create a `HistorySourcePass` node to define the type of texture and initial state. This node's output pin will give access to the previous frame's content.
Next, create an `ezHistoryTargetPass`. It's input pin exposes the same texture as provided by the source node but allows you to write to by connecting the input pin to another pass that produces the image that you want to carry to the next frame. To connect an ezHistoryTargetPass to its counterpart you need to set it's "SourcePassName" property to the name of the ezHistorySourcePass you want to match.
As both nodes expose the same texture, special care has to be taken that it's not used as input and output of another pass at the same time. In those cases, add a `ezCopyTexturePass` to break up invalid state.

Added a very simple blend pass that blends the two input pins by an exposed BlendFactor.

Fixed Vulkan Copy Texture going out of bounds when copying textures with mip maps.

---------

Co-authored-by: ezEngine Automation <ezEngineProject@gmail.com>