Added component to trigger scene transitions (#1363)
Adds a component that makes it easy to switch between levels.

ezSceneTransitionComponent is a component that either reacts to trigger events, or can be invoked via script to make the game state load into another scene.
It can also trigger (or cancel) a preload, and forward relative player positions to the new scene, to make it possible to build levels with very natural transitions. You can also select at which start point a player should spawn, in case the target level has multiple.

There is a new demo scene that shows different possibilities. The video below shows the setup for preloading a scene with one trigger and then entering the scene through one or another entry point.