Add support for cross-compiling for the SteamRT Sniper (#1391)
This adds support for cross-compiling the ezEngine runtime for the SteamRT Sniper. This includes a CMakeUserPresets.json setup to do so. This allows ezEngine to run on the SteamRT, which means it can be shipped as a Linux game on Steam and requires only one build to be done and will then run on all Linux distributions Steam supports. With this there is no need to do a separate ezEngine build for each Linux distribution.