Refactor of platform specific includes (#1414)
If a library has a folder "Platform/XYZ" (with XYZ being the target platform name), then this is automatically added as a public include directory. This allows to do #include<Feature.h> where "Feature.h" will redirect to a platform specific header.

This means, that we don't need to have the #ifdef PLATFORM_XYZ cascades in many places anymore, and adding a platform becomes much easier, because you don't need to modify existing code.

If code for a platform is missing, the #include fails at compile time, and no "#else error" is needed.

* Moved many platform specific files around to clean them up and automatically include them.
* Removed now unnecessary #ifdef sections.
* FileServe now automatically loads all available shader compiler plugins

There's still quite a lot more to do, to separate platform code from the main code, but that may come later.

This is part of #1119.