Fix monitors in testing chambers main scene
Fixes #1410
The vignette texture was not set, which is also the default for the tonemap pass. With the fallback resources the fallback color was changed to black instead of white (which is what DX11 produces when sampling unbound texture slots).
I've changed the render pass / pipeline instead of messing with the fallback resources as these are not supposed to be sampled in any case.

Also fixed a wrong pass after the scene was renamed.