Vulkan Descriptor Set Optimizations
# Vulkan Descriptor Set
* Descriptor sets are now cached within a frame.\
To be able to cache descriptor sets, we not only need the bind group description but also the currently used dynamic uniform buffers used by transient constant buffers.
All constant buffers in EZ are marked as dynamic in the layout so we need to provide offsets for each slot, no matter if a normal or transient constant buffer is bound to a slot.
* Any time a transient constant buffer is updated, `m_bDynamicOffsetsDirty` is set and before drawing we need to update all dynamic uniform buffer offsets in each bind group. If the current dynamic uniform buffer is exhausted, we need to create a new descriptor set for each affected bind group. See return value of `UpdateDynamicUniformBufferOffsets`.
* Whenever a bind group or it's offsets change, all corresponding descriptor sets need to be updated via `bindDescriptorSets`.
* Calls to `updateDescriptorSets` are batched if multiple descriptor sets need to be created before a draw call. No further batching can be done without `VK_DESCRIPTOR_BINDING_UPDATE_AFTER_BIND_BIT_EXT` which requires Vulkan 1.2.
* Added some statistics.