Fixed #1479: Added option to pause and step simulation in editor
Added a pause button to the editor. When paused, the button turns into a "step" button, that allows you to advance the simulation by one frame. Click "play" to resume regular simulation.

Also fixed that the world-time didn't work for decal materials. And switch the caustics decal material to use the world time.