Stutter fix in editor play the game
Previously the editor rendered in two places:
1. `ezEngineProcessGameApplication::EventHandlerIPC` if a `ezSyncWithProcessMsgToEngine` message was received.
2. At the end of `ezEngineProcessGameApplication::Run`.
This meant that in one frame loop multiple calls to `ezGameApplicationBase::RunOneFrame` could happen depending on what the editor wanted to render and what the running play the game loop wanted to render. This caused stuttering as the frames were very unbalanced in terms of workload.

Now, to ensure that for each `ezSyncWithProcessMsgToEngine` message a frame is rendered, a new feature was added to cancel out of message processing via the new event message field `m_bInterruptMessageProcessing`. For each `ezSyncWithProcessMsgToEngine` message, we set `m_bInterruptMessageProcessing` to true so the message processing is interrupted and we end up at `ezEngineProcessGameApplication::Run` where  all the rendering now takes place. This ensures that rendering does not happen outside of the normal game loop.

## Misc
Fixed a crash in ImGui when closing the play the game window of the sample game project. There was no code path to ensure that at the end of the frame the ImGui frame was ended. So when the play the game window was closed, ImGui was shut down but the destructor asserted that a frame is still in progress.