Make standard app input configurable through ezGameState
ezGameApplication has some standard input setup.
For example this implements that ESC always closes the application,
and F5 shows the FPS counter, etc.
This is now configurable. The ezGameState makes the call to set this up,
and custom game states can this decide not to do this.
The RTS sample now uses this to deactivate the ESC key.
In the future it may then show an ingame menu instead.