Removed VERTEX_SHADER_RENDER_TARGET_ARRAY_INDEX shader permutation
Geometry shader fallback for stereo rendering was only relevant for DX11.0 cards that did not support the extension but did support geomertry shaders. Every desktop card supports the extension and on mobile neither geometry shaders nor the array index extension are supported so stereo rendering has to be done with different extensions anyways. The capability remains as it currently defines whether stereo rendering is supported.